Split test_withings_routes: Move tests for multiple notification webhook calls to a separate file.
